Feat : Backend - 이미지 변형 기능(코드 리뷰 반영)
- SecurityConfig : comfy webhook 경로 permitAll
- ApiLog : RunComfy 측 request ID 기록을 위한 속성 + 마이그레이션 스크립트 추가(V7__)
- ComfyWebhookService : 인터페이스 계층 제거, webhook payload <-> ApiLog.externalRequestId 검증 추가, DB 업데이트 시점에서 lock 획득하도록 수정(finalizeSuccess)
