+++ backend-spring/today-store/src/main/resources/db/migration/V7__add_external_request_id_to_api_logs.sql
@@ -1,0 +1,2 @@
+-- V7: Add external_request_id to api_logs table
+ALTER TABLE api_logs ADD COLUMN external_request_id VARCHAR(100);
